raftstore: Use engine traits
