+++ tests/failpoints/mod.rs
@@ -4,6 +4,12 @@
 
 #[macro_use]
 extern crate slog_global;
+#[macro_use]
+extern crate slog_global;
+#[macro_use(box_err, box_try)]
+extern crate tikv_util;
+#[macro_use(fail_point)]
+extern crate fail;
 
 mod cases;
